league_mgt/requirements.txt
asgiref==3.2.3
Django==3.0.3
djangorestframework==3.11.0
pyton==3.8.7
sqlparse==0.3.0
